Bugfix in IndexTriangle where default template argument did not match that of IndexedTriMesh
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2869 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/IndexedTriangle.hpp
@@ -34,7 +34,7 @@ namespace geometry {
 	 * vertices in an array typically used with the IndexedTriMesh  class.
 	 * the indice type (size) is templated.
 	 */
-	template<class T = uint16_t>
+	template<class T = uint32_t>
 	class IndexedTriangle {
 	private:
 		//! @brief indices to the vertices of the triangle
